fix(scan): Delete pre-registry image tag
When doing pre-registry scans, the action tags the image for the
pre-registry host. Previously the tags were not cleaned up, which
could cause useless tags and images to build up on the build host.

This change makes the action delete the pre-registry tag when it
is no longer needed so that images can be garbage-collected more
easily.