fix: do not modify image reference
Previously we would default the image registry to `docker.io` if it was
not specified, but this is not correct in some instances. This code does
not have enough information to apply the defaults, so leave the values
as they are and let the scanner (which does have enough information to
make a proper decision) fill in the blanks.